Fixed Number type from float to double according to the Praat's implementation.

+++ textgrid.hpp
@@ -85,7 +85,7 @@ class TextGridVisitor {
   virtual void Visit(const Point& point) = 0;
 };
 
-typedef float Number;
+typedef double Number;
 
 struct Point {
   Point() = default;
